[#24] reorganizando rotas
