+++ app/Http/Controllers/Api/Paciente/Historico/HistoricoController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api\Historico;
+namespace App\Http\Controllers\Api\Paciente\Historico;
 
 use App\Api\ErrorMessage;
 use App\Models\Historico;
@@ -20,26 +20,25 @@ class HistoricoController extends Controller
         $this->historico = $historico;
     }
 
-    public function show($paciente_id)
+    public function show(int $pacienteId)
     {
         try {
 
-            if (!Paciente::find($paciente_id)) {
+            if (!Paciente::find($pacienteId)) {
                 return response()->json(['message' => 'Paciente não existe'], 400);
             }
 
-            $historico = $this->historico->where('paciente_id', $paciente_id)->first();
+            $historico = $this->historico->where('paciente_id', $pacienteId)->first();
 
             if (!isset($historico)) {
                 return response()->json(['message' => 'Paciente não possui histórico cadastrado'], 400);
             }
-            
+
             $drogas = $this->historico->find($historico->id)->drogas()->get();
 
             $historico['drogas'] = $drogas;
 
             return response()->json($historico);
-
         } catch (\Exception $e) {
             $message = new ErrorMessage($e->getMessage());
             return response()->json($message->getMessage(), 500);
@@ -62,15 +61,17 @@ class HistoricoController extends Controller
                 return response()->json(['message' => 'Paciente já possui historico'], 400);
             }
 
-            if ($request->has('situacao_uso_drogas_id')
-                && !SituacaoUsoDrogas::find($request->get('situacao_uso_drogas_id'))) {
+            if (
+                $request->has('situacao_uso_drogas_id')
+                && !SituacaoUsoDrogas::find($request->get('situacao_uso_drogas_id'))
+            ) {
                 return response()->json(['message' => 'Situação não existe'], 400);
             }
 
             $historico = $this->historico->create($data);
 
-            if(isset($data['drogas']) && count($data['drogas'])) {
-                foreach($data['drogas'] as $droga_id) {
+            if (isset($data['drogas']) && count($data['drogas'])) {
+                foreach ($data['drogas'] as $droga_id) {
                     if (!Droga::find($droga_id)) {
                         return response()->json(['message' => "Droga de id $droga_id não existe"], 400);
                     }
@@ -80,7 +81,6 @@ class HistoricoController extends Controller
             }
 
             return response()->json($historico);
-
         } catch (\Exception $e) {
             $message = new ErrorMessage($e->getMessage());
             return response()->json($message->getMessage(), 500);
@@ -93,14 +93,13 @@ class HistoricoController extends Controller
         try {
 
             $data = $request->all();
-            
+
             $historico = $this->historico->findOrFail($id);
             $historico->update($data);
 
             return response()->json([
                 'message' => 'Historico atualizado com sucesso'
             ]);
-
         } catch (\Exception $e) {
             $message = new ErrorMessage($e->getMessage());
             return response()->json($message->getMessage(), 500);
